[CAP-005] 요약상세보기 ViewModel 파일로 분리
